moving get games logic to api
